feat: undo example folder restructuration
