+++ example/src/main/java/com/elevenpaths/almaraz/example/User.java
@@ -2,7 +2,7 @@
  * Copyright (c) Telefonica I+D. All rights reserved.
  */
 
-package com.elevenpaths.almaraz.example.model;
+package com.elevenpaths.almaraz.example;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
